AG/ROBOI/nareshCodes2021-roboi-edge-service-ddf3c8e7bdc2/nareshCodes2021-roboi-edge-service-ddf3c8e7bdc2/uploader/requirements.txt
google-cloud-storage==2.14.0
pyyaml
